Rearrange stimcirq
