+++ glue/cirq/stimcirq/__init__.py
@@ -1,5 +1,6 @@
-from ._stim_sampler import StimSampler, cirq_circuit_to_stim_circuit
-from ._stim_to_cirq_circuit_conversion import stim_circuit_to_cirq_circuit, MeasureAndOrReset
+from ._cirq_to_stim import cirq_circuit_to_stim_circuit
+from ._stim_sampler import StimSampler
+from ._stim_to_cirq import stim_circuit_to_cirq_circuit, MeasureAndOrReset
 
 # Workaround for doctest not searching imported objects.
 __test__ = {
