+++ glue/cirq/stimcirq/_stim_to_cirq.py
